OLMIS-1905 Changed period dates and removed some clutter from features
The reason for changing the period dates is that the tests update one
of the periods to the current date. This leads to a situation that
we have two periods overlapping, which is illegal. I've changed the
dates to 2016 to avoid that. I've also removed some unnecessary steps
for emergency requisition tests - initiating and submitting regular
requisition should not be needed there.